# MapSeq To XML Performance Improvement: >4000x speedup