@tobiade 2024-03-29
I've also just realised this issue was highlighted in #70 but looks like we missed out getting a fix for `MapSeq` 😄