@clbanning 2024-04-03
> I've also just realised this issue was highlighted in #70 but looks like we missed out getting a fix for `MapSeq` 😄

That was intentional.  MapSeq was a special request and not viewed as a widely used option at the time; getting the core functionality performing better was believed to, by far, have the most impact for the package.